machine_learning_study/requirements.txt
numpy
scikit-surprise